+++ src/test/java/com/gonghak98/v2/abeek/major/LabMajorTest.java
@@ -16,7 +16,7 @@ class LabMajorTest {
     @ParameterizedTest
     @CsvSource(value = {"5611, 디지털논리회로", "9658, 전기회로실험", "8076, 기초광학및실험", "-1, 전자소자공정실험", "9666, 마이크로컴퓨터실험"})
     @DisplayName("지정된 실험교과목을 1과목 이상 이수하면, 실험 교과목 이수조건을 만족한다.")
-    void 한_개_과목이상_이수(int courseId, String studentCourseName) {
+    void 한_개_과목이상_이수(Long courseId, String studentCourseName) {
         //given
         List<CompletedCourse> studentCourses = 랜덤이름을_가진_기이수_과목_특정_개수만큼_생성(5);
         studentCourses.add(CompletedCourse.builder().id(courseId).name(studentCourseName).build());
